Modularizing ResetPassword.
