+++ src/account-settings/reset-password/data/actions.js
@@ -1,0 +1,22 @@
+import { utils } from '../../../common';
+
+const { AsyncActionType } = utils;
+
+export const RESET_PASSWORD = new AsyncActionType('ACCOUNT_SETTINGS', 'RESET_PASSWORD');
+
+export const resetPassword = email => ({
+  type: RESET_PASSWORD.BASE,
+  payload: { email },
+});
+
+export const resetPasswordBegin = () => ({
+  type: RESET_PASSWORD.BEGIN,
+});
+
+export const resetPasswordSuccess = () => ({
+  type: RESET_PASSWORD.SUCCESS,
+});
+
+export const resetPasswordReset = () => ({
+  type: RESET_PASSWORD.RESET,
+});
