+++ src/account-settings/reset-password/index.js
@@ -1,0 +1,5 @@
+export { default } from './ResetPassword';
+export { default as reducer } from './data/reducers';
+export { RESET_PASSWORD } from './data/actions';
+export { default as saga } from './data/sagas';
+export { configureService } from './data/service';
