+++ src/account-settings/reset-password/messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from '@edx/frontend-i18n';
+
+const messages = defineMessages({
+  'account.settings.editable.field.password.reset.button': {
+    id: 'account.settings.editable.field.password.reset.button',
+    defaultMessage: 'Reset Password',
+    description: 'The password reset button in account settings',
+  },
+});
+
+export default messages;
